Make Interwiki macro replace spaces & titlecase
This change makes the Interwiki macro construct the path value for wiki
URLs by replacing any spaces in the article title with underscores, and
uppercasing the first letter of the path.

Otherwise, without this change, the macro constructs wiki URLs that
contain spaces — which results in MDN having more than 100 URLs with
spaces; and because the URL spec disallows spaces in URLs, all the MDN
articles with those space-containing URLs are invalid.

Moreover, given a URL with spaces, MediaWiki anyway redirects that to
the corresponding URL which has the spaces replaced by underscores. So
this change also prevents an unnecessary/wasteful redirect.

Further, given a URL with a path that begins with a lowercase letter,
MediaWiki redirects that to the corresponding URL which has the first
letter of the path in uppercase.

And MDN has 190 Interwiki macro calls with values that don’t begin with
an uppercase letter, so without the part of this change which uppercases
the first letter of the path, the result is that we have 190 links that
cause unnecessary/wasteful redirects each time anyone follows them.