Drop cases adding "closed" attribute to <details>
This change updates all sidebar code that adds the <details> element
with a "closed" attribute. Per the HTML spec, the <details> element has
no "closed" attribute; so every MDN article which has <details closed> in
a sidebar becomes invalid. The "closed" attribute is anyway unnecessary;
https://html.spec.whatwg.org/multipage/#attr-details-open says:

> The "open" content attribute is a boolean attribute. If present, it
> indicates that both the summary and the additional information is to be
> shown to the user. If the attribute is absent, only the summary is to be
> shown.

So in other words, it’s sufficient to just omit the "open" attribute —
because doing so has exactly the same meaning as "closed".