Allow URL or pathname for "yarn content preview"
This change allows you to give "yarn content preview" an MDN URL or
local pathname (with or without the "index.html" part):

    yarn content preview files/en-us/web/api/idbindex
    yarn content preview files/en-us/web/api/idbindex/index.html
    yarn content preview https://developer.mozilla.org/en-US/docs/Web/API/IDBIndex
    yarn content preview https://developer.mozilla.org/docs/Web/API/IDBIndex

The local-pathname support allows you to yari-view a given article by
using the tab key and your shell’s pathname-completion mechanism to
complete a local directory path in your mdn/content clone.

It also facilitates using the complete paths copied from, for example,
the "Files changed" tab of the GitHub pull-request UI, or from the
output of "git log --name-only" or "git diff --name-only" — all of which
show the file pathnames in the form:

    files/en-us/web/javascript/reference/global_objects/date/index.html

This change allows you to paste that full pathname as-is into your shell
as the argument for "yarn content preview".

The URL support allows you to preview the local equivalent of whatever
production https://developer.mozilla.org URL you give.