Make Yari handle .jpg files (not just .jpeg)
This change makes the Yari backend correctly handle .jpg files.

Otherwise, without this change, the Yari (Express) backend responds with
a 404 to any requests for .jpg resources; for example:
http://localhost:5000/en-US/docs/Web/API/Window/devicePixelRatio/devicepixelration_diff.jpg

Files with .jpeg extensions are already correctly handled, even without
this change; so the patch here is a minor change to update an existing
regex pattern in the code to make it match ".jpg" as well as ".jpeg".