Fix broken links to in-page sections (#2383) (#2385)
When being rendered, the header IDs are converted to lowercase, whereas
the [anchor macro] does not convert its argument to lowercase. This
causes in-page links to be broken on pages that pass IDs containing
capital letters to the anchor macro. This commit fixes it by converting
the argument of the anchor macro to lowercase, with the help of
[string.toLower function].

[anchor macro]: kumascript/macros/anch.ejs
[string.toLower function]: kumascript/src/api/string.js