Add release date for support statements with multiple versions
The code that added in the release date did not account for support
statements that involve multiple releases (for example, a version with
flags and then a version with full support), so these entries did not
have a release date tooltip in the compatibility table.

This commit extends the release date addition to annotate the release
date for all versions in a support statement.