fix: correct flaws fixing logic
This commit fix the potential problem of document flaws fixing.

In current logic, if the fixFixableFlaws function is not executed in the
verbose mode, the program will still update the document even if it's in
dry run mode.

If a program is executed in dry run mode, it should show what the
program is attempting to perform. So, we don't need to check whether it
is verbose or not here.

Also, although the current default setting (build/constants.js) of
FIX_FLAWS_VERBOSE is true. The flaws fixing logic should not depend on
it.