fix: cli flaws command logic & minor refactorings
This commit modified flaws fixing logic in the cli and do some minor
refactorings.

1. In current logic, if any flaws is found in document, the cli tool
will ask user whether user wants to fix the flaws or not. However, not
all flaws are fixable. Suppose there's one document contains only one
non-fixable flaw and user ask the program to fix it, the user will not
get any feedback currently. The program should check if one flaw is
fixable before asking user whether to fix it or not.

2. Since yari requires Node.js (>= 12.0.0), I used Array.prototype.flat,
which is available since 11.0.0, to simplify current flaws counting
logic.

3. Rephrase "Found no fixable flaws!" to "✓ All seems fine".