strip prettier-ignore from source with AST positions
This marks the 3rd change of how we strip those tags, and a pendulum swing in an opposite direction of the last one.
The 1st version simply string replaced all the prettier-ignore tags, which could lead to invalid ASTs (imagine a prettier ignore as a first item in a listItem, which would turn into an invalidly-empty first item space).
The 2nd (aka previous) version simply dropped the HTML AST nodes which represented the ignore, but we then used mdast's stringify to turn it back into source which would mess up some of prettier's work (Hamish found that there was too much spacing around list-items). Now we could not throw it back into prettier as the input lost its prettier-ignore tags, which would lead to it wrecking our code-blocks.
Hence this 3rd version is a mix of the two!