Make Specifications macro generate spec URL for ARIA docs based on title
This change makes the Specifications macro generate a spec URL for all
articles for ARIA attributes based on the article title, rather than by
doing a BCD query.

This makes it possible to just put a {{Specifications}} macro call in
all the ARIA docs, and have it automatically do the right thing.

Otherwise, without this change, we’d need to add data to BCD for all the
ARIA attributes. It’s not clear that we’d ever actually want to do that —
but anyway, this change is forward-compatible in that if we ever did
add all the ARIA attributes to BCD, this change won’t break anything
(instead it would end up just being redundant, and so we could remove it
at that time).