fix(search): allow submitting non-empty query via button
Previously, the search form could only be submitted by pressing
the Enter key while having focus in the input field, and any
regular submits via the submit button were prevented.

This was meant to prevent submitting the form without a query term,
but it also prevented submitting it *with* a non-empty query term.