chore(markdown): copy defaultHandler from mdast-util-to-hast
To create a custom handler wrapping tables, we need the default
handler for tables, which is part of the `mdast-util-to-hast` package.

However, that package is an ES module, and we cannot currently import
exports from ES modules.

As a ugly workaround, we copy the required parts from that library,
and from another one (`unist-util-position`) that it relies on.