fix:themes - switch to theme drop down from toggle
This removes the current theme toggle and replaces it with the theme dropdown.
This then also exposes the ability for a reader to set the theme to match the os default.

NOTE: This fixes an issue that is logged in the yari private repo so, I cannot link it here.