Allow the browser-compat frontmatter value to be an array
We currently have some existing documents that use multiple
{{Compat(...)}} macros within the same document — because without the
browser-compat frontmatter key being allowed to contain an array,
there’s no other way to have multiple BCD tables in the same document
except by putting multiple {{Compat(...)}} macros in the document.

So, allowing the browser-compat value to contain an array eliminates the
need for authors to put multiple {{Compat(...)}} macros in a document.

And reducing the need for authors to use the {{Compat(...)}} macro at all
is also in line with our general de-macro-ization plans — in that it takes
us a big step closer to completely eliminating the {{Compat}} altogether.