Allow authors to omit the {{Compat}} and {{Specifications}} macros
For the normal case where a {{Compat}} or {{Specifications}} macro has
no arguments, this change allows authors to completely omit the macro.

That is, this change allows authors to just do the following:

  ## Specifications

  ## Browser compatibility

... rather than needing to do this:

  ## Specifications

  {{Specifications}}

  ## Browser compatibility

  {{Compat}}

This change works by directly passing the frontmatter metadata down to
the code that builds the contents of the “Browser compatibility” and
“Specifications” sections — rather than needing for it to be indirectly
passed by the {{Compat}} and {{Specifications}} macros.

This change is also in line with our general de-macro-ization plans — in
that it takes us a big step closer to completely eliminating the
{{Compat}} and {{Specifications}} macros.