fix(css): set scroll-behavior unconditionally
Previously, the scroll-behavior was only set when an element had focus,
which can be very short, e.g. when closing the user menu or clicking a
link. In Firefox this caused a reflow, and interactive examples flickered.

See: https://bugzilla.mozilla.org/show_bug.cgi?id=1764655

Now, we set the scroll-behavior unconditionally, and this avoids both
the reflow/flicker, but also fixes the scroll-behavior not working.

Co-authored-by: Daniel Holbert <dholbert@mozilla.com>