Make Specifications work as expected with multiple BCD queries
Without this change, when there are multiple BCD features, the
Specifications section won’t get populated as expected.

I tested this change and I believe it’ll work correctly  for all cases.

(And I think we can switch to Object.values rather than Object.entries,
because we only need the values, and don’t really need to do anything
with the keys.)