Further update BCD table features population to handle multiple BCD queries
Some features — e.g., css.properties.justify-content — have no
compat data themselves but have subfeatures with compat data.
In the case where we’re processing a browser-compat value which is an
array — that is, multiple top-level features in one table (in contrast
to one feature and its subfeatures) — unless we explicitly check for the
__compat-less-feature-with-subfeatures case, and handle it, we can end
up with rows for expected features being entirely missing from the
rendered BCD table.

So this change adds handling for that case.