Avoid data-clobbering and duplicate features in multi-BCD-queries case
In _getBCDData(queryString), the queryString value) can be, e.g.,
"html.elements.link.rel,html.elements.a.rel,html.elements.area.rel" for
the case where we’re processing multiple BCD queries. So as the key in
the data structure we create there and pass around to other parts of the
code, to ensure the key is unique, we use the full BCD query/feature
name rather than some portion of it.

Otherwise, if we use, e.g., just the part after the last dot, then for
the example above, that ends up being just 'rel' for all three 'query'
values from that full queryString value. And because 'rel' isn’t unique,
if we used it as the key in our internal data structure, then in each
iteration through the 'queryString', we’d end up setting a new value for
the 'rel' key — and clobbering whatever value the previous iteration set.

And then — to make things very clear — we pass around that key/property
to other parts of the code using the variable 'query' to identify it
(rather than 'name').

And then later, for the case when we’re processing multiple BCD 'query'
instances (BCD queries/features) in the part of the code that outputs
the rendered BCD table — where a shortened identifier is needed — we can
end up with duplicate identifiers being displayed unless we further
qualify them.

The query string "html.elements.label.for,html.elements.output.for", for
example, would result in a rendered BCD table that has just 'for' as the
feature name in each row. So in this case, we take the last _two_ parts
of the query/feature name — that is, e.g., 'label.for' and 'output.for'
— and display those in the rendered table.