Add HTML “title” attribute to BCD table row headers
This change adds an HTML “title” attribute too all BCD table row
headers, so that users can hover to see the full BCD feature name.

Otherwise, without this change, for some multiple-BCD-queries cases,
the displayed names of the features are the same; for example, consider
"css.properties.width.fit-content_function,css.properties.grid-template-columns.fit-content"
For the case of those two features, the displayed feature name for both
is just “fit-content()” — so there’s no way for users to distinguish
which row applies to which feature.

But with this change, users can hover the table row headers to see which
is for which.