chore: build spotlight contributors for all locales
Currently, the spotlight contributor pages and assets are only generated for the `en-us` folder.
This causes a 404 if people from other locales tries to access the spotlight contributor page.
In https://github.com/mdn/yari/pull/6430 there was a temporary fix to hard code the locale into the link.
Perhaps we should take this approach for the longer term?