Make Specifications work as expected with multiple BCD queries (#6465)
* Make Specifications work as expected with multiple BCD queries

Without this change, if there are multiple BCD features (array value in
browser-compat), the Specifications section isn’t populated as expected.

* Further update Specifications population to handle multiple BCD queries

Some features — e.g., css.properties.justify-content — have no compat
data themselves but have subfeatures with compat data. In the case
where we’re processing a browser-compat value which is an array, unless
we explicitly check for the __compat-less-feature-with-subfeatures case,
and handle it, we can end up with expected specifications being entirely
missing from the rendered Specifications section.

So this change adds handling for that case.