fix(macros/ListSubpagesForSidebar): list parent + subpages, if requested
Previously, setting the third parameter ($2) would only list the parent
itself, but not its subpages.

In theory, `page.subpagesExpand()` is broken, because it calls
`info.getChildren()` which doesn't have a `depth` parameter and
just returns *either* the parent *or* its subpages.

However, `page.subpagesExpand()` is used in several places, so
"fixing" that helper function is risky.