fix(esm): update tsconfigs, fix files
fix base tsconfig to use "NodeNext" module and moduleResolution: these
settings integrate with node's native esm support.

fix the resulting typescript errors from this change.

downgrade the @types/imagemin package to match our installed version

fix incorrect types on "front-matter" package exports with the
type-fixes/front-matter.ts file and "paths" tsconfig compilerOption

decouple client/pwa/ssr tsconfig from base tsconfig: they all run
through webpack - it takes care of module resolution - so we set
moduleResolution to "Node" as that's the form it expects it to be in.
we leave module as "ES2020" as this is (apparently) required for
tree-shaking to work.