feat(sidebars): support status meta for badges (#7828)
Specific status metadata is replacing tag metadata for experimental,
non-standard, and deprecated. Initial work to expose the metadata 
was done in #7801. This is follow on work to use the status metadata 
as well as the tags. There will be a follow on PR to remove the tag 
parsing once all the content has been updated.

Co-authored-by: Will Bamberg <will@bootbonnet.ca>
Co-authored-by: Claas Augner <caugner@mozilla.com>