Remove "attr-" from HTML Element attribute links
The prefix "attr-" was currently added to URL fragment to locate the HTML element attribute on a page. However, it is impossible to follow the URL to locate its location as the "id" attribute lacks this prefix.

For example, I can not locate "colspan" section by clicking the "colspan" link on the "[table](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/table)" page, as the URL contains the incorrect fragment "#attr-colspan". The correct fragment should be "#colspan".