fix(a11y): fix tab order reset in the header (#9851)
fix(tab): fix tab order reset in the header

When tabbing through the header reaching the search input and
pressing tab again will reset the focus to the html document.

This happened because the next item would be the cancel input
button. Which will be removed via display one and then we start
from the the beginning.