fix(plus): scroll to anchor hook
When navigating to a page via anchor link, scroll to the anchor via JS.
We need this for non SSRed pages (mainly plus).

Correct scroll-margins is only supported Firefox (via :target) fully.
For some reason Blink and WebKit seem to not set :target consitently.
To work around this we set scroll-margin for section[id], too.