feat(experiment): rewrite Web/API page titles (#10926)
Problem: We identified a drop in search engine impressions in April 2023, which coincides
with the Web/API retitle project (https://github.com/mdn/mdn/issues/284), and we cannot
rule out that there is a causal relationship.

Solution: Run an experiment with 500 randomly sampled Web/API pages (of 2719 affected),
partially reverting  the change to their `<title>` (but not to the `<h1>`) by essentially
replacing `Foo: bar property` with `Foo.bar property`.