fix(macros/ListSubpages): list subpages via default locale
Now, the `wiki.tree()` method will include all the subpages even if the page is not localized. This This improves the experience for non-English users.

Additional changes:
The type hint has been added to some methods to improve the development experience.
The built-in sorting methods has been used to sort numerically.